fix the chan 0 rollover
git-svn-id: svn://orca.physics.unc.edu/Orca@10525 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/Vme/CAEN/CV830/ORCV830Model.h
@@ -125,7 +125,7 @@ enum {
 		NSArray* dataTakers;	//cache of data takers.
 		unsigned long numEnabledChannels;
         unsigned long    lastChan0Count;
-        unsigned long    chan0RollOverCount;
+        unsigned long long    chan0RollOverCount;
     
 }
